+++ github4s/js/src/main/scala/github4s/util/URLEncoder.scala
@@ -23,6 +23,7 @@ package github4s.util
 
 object URLEncoder {
 
-  def encode(string: String): String = fr.hmil.roshttp.util.Utils.encodeQueryString(string)
+  def encode(string: String): String =
+    fr.hmil.roshttp.util.Utils.encodeQueryString(string)
 
 }