+++ github4s/js/src/test/scala/github4s/utils/integration/GHUsersSpec.scala
@@ -32,10 +32,12 @@ import github4s.js.Implicits._
 
 class GHUsersSpec extends AsyncFlatSpec with Matchers with TestUtils {
 
-  override implicit val executionContext = scala.concurrent.ExecutionContext.Implicits.global
+  override implicit val executionContext =
+    scala.concurrent.ExecutionContext.Implicits.global
 
   "Users >> Get" should "return the expected login for a valid username" in {
-    val response = Github(accessToken).users.get(validUsername).execFuture(headerUserAgent)
+    val response =
+      Github(accessToken).users.get(validUsername).execFuture(headerUserAgent)
 
     testFutureIsRight[User](response, { r =>
       r.result.login shouldBe validUsername
@@ -44,20 +46,25 @@ class GHUsersSpec extends AsyncFlatSpec with Matchers with TestUtils {
   }
 
   it should "return error on Left for invalid username" in {
-    val response = Github(accessToken).users.get(invalidUsername).execFuture(headerUserAgent)
+    val response = Github(accessToken).users
+      .get(invalidUsername)
+      .execFuture(headerUserAgent)
 
     testFutureIsLeft(response)
   }
 
   "Users >> GetAuth" should "return error on Left when no accessToken is provided" in {
-    val response = Github().users.getAuth.exec[Future, SimpleHttpResponse](headerUserAgent)
+    val response =
+      Github().users.getAuth.exec[Future, SimpleHttpResponse](headerUserAgent)
 
     testFutureIsLeft(response)
   }
 
   "Users >> GetUsers" should "return users for a valid since value" in {
     val response =
-      Github(accessToken).users.getUsers(validSinceInt).execFuture(headerUserAgent)
+      Github(accessToken).users
+        .getUsers(validSinceInt)
+        .execFuture(headerUserAgent)
 
     testFutureIsRight[List[User]](response, { r =>
       r.result.nonEmpty shouldBe true
@@ -67,7 +74,9 @@ class GHUsersSpec extends AsyncFlatSpec with Matchers with TestUtils {
 
   it should "return an empty list when a invalid since value is provided" in {
     val response =
-      Github(accessToken).users.getUsers(invalidSinceInt).execFuture(headerUserAgent)
+      Github(accessToken).users
+        .getUsers(invalidSinceInt)
+        .execFuture(headerUserAgent)
 
     testFutureIsRight[List[User]](response, { r =>
       r.result.isEmpty shouldBe true