+++ scalaz/src/main/scala/github4s/scalaz/Implicits.scala
@@ -38,25 +38,28 @@ object implicits extends HttpRequestBuilderExtensionJVM {
     override def capture[A](a: ⇒ A): Task[A] = Task.now(a)
   }
 
-  implicit def g4sTaskMonadError: MonadError[Task, Throwable] = new MonadError[Task, Throwable] {
+  implicit def g4sTaskMonadError: MonadError[Task, Throwable] =
+    new MonadError[Task, Throwable] {
 
-    override def pure[A](x: A): Task[A] = Task.now(x)
+      override def pure[A](x: A): Task[A] = Task.now(x)
 
-    override def map[A, B](fa: Task[A])(f: A ⇒ B): Task[B] = fa.map(f)
+      override def map[A, B](fa: Task[A])(f: A ⇒ B): Task[B] = fa.map(f)
 
-    override def flatMap[A, B](fa: Task[A])(ff: A ⇒ Task[B]): Task[B] = fa.flatMap(ff)
+      override def flatMap[A, B](fa: Task[A])(ff: A ⇒ Task[B]): Task[B] =
+        fa.flatMap(ff)
 
-    override def tailRecM[A, B](a: A)(f: A ⇒ Task[Either[A, B]]): Task[B] =
-      Task.tailrecM((a: A) ⇒ f(a) map (t ⇒ toScalazDisjunction(t)))(a)
+      override def tailRecM[A, B](a: A)(f: A ⇒ Task[Either[A, B]]): Task[B] =
+        Task.tailrecM((a: A) ⇒ f(a) map (t ⇒ toScalazDisjunction(t)))(a)
 
-    override def raiseError[A](e: Throwable): Task[A] = Task.fail(e)
+      override def raiseError[A](e: Throwable): Task[A] = Task.fail(e)
 
-    override def handleErrorWith[A](fa: Task[A])(f: Throwable ⇒ Task[A]): Task[A] =
-      fa.handleWith({ case x ⇒ f(x) })
+      override def handleErrorWith[A](fa: Task[A])(f: Throwable ⇒ Task[A]): Task[A] =
+        fa.handleWith({ case x ⇒ f(x) })
 
-  }
+    }
 
-  implicit val intInstanceTaskScalaJ = new Interpreters[Task, HttpResponse[String]]
+  implicit val intInstanceTaskScalaJ =
+    new Interpreters[Task, HttpResponse[String]]
 
   private[this] def toScalazDisjunction[A, B](disj: Either[A, B]): A \/ B =
     disj.fold(l ⇒ -\/(l), r ⇒ \/-(r))